+++ apps/io-wallet-user-func/src/infra/http/handlers/__test__/get-current-wallet-instance.status.spec.ts
@@ -20,7 +20,6 @@ describe("GetCurrentWalletInstanceStatusHandler", () => {
   const walletInstanceRepository: WalletInstanceRepository = {
     batchPatch: () => TE.left(new Error("not implemented")),
     get: () => TE.left(new Error("not implemented")),
-    getAllByUserId: () => TE.left(new Error("not implemented")),
     getLastByUserId: () =>
       TE.right(
         O.some({
@@ -37,6 +36,7 @@ describe("GetCurrentWalletInstanceStatusHandler", () => {
           userId: "AAA" as FiscalCode,
         }),
       ),
+    getValidByUserIdExcludingOne: () => TE.left(new Error("not implemented")),
     insert: () => TE.left(new Error("not implemented")),
   };
 
@@ -136,8 +136,8 @@ describe("GetCurrentWalletInstanceStatusHandler", () => {
     const walletInstanceRepository: WalletInstanceRepository = {
       batchPatch: () => TE.left(new Error("not implemented")),
       get: () => TE.left(new Error("not implemented")),
-      getAllByUserId: () => TE.left(new Error("not implemented")),
       getLastByUserId: () => TE.right(O.none),
+      getValidByUserIdExcludingOne: () => TE.left(new Error("not implemented")),
       insert: () => TE.left(new Error("not implemented")),
     };
     const handler = GetCurrentWalletInstanceStatusHandler({
@@ -164,8 +164,9 @@ describe("GetCurrentWalletInstanceStatusHandler", () => {
       {
         batchPatch: () => TE.left(new Error("not implemented")),
         get: () => TE.left(new Error("not implemented")),
-        getAllByUserId: () => TE.left(new Error("not implemented")),
         getLastByUserId: () => TE.left(new Error("failed on getLastByUserId!")),
+        getValidByUserIdExcludingOne: () =>
+          TE.left(new Error("not implemented")),
         insert: () => TE.left(new Error("not implemented")),
       };
     const handler = GetCurrentWalletInstanceStatusHandler({
@@ -193,8 +194,9 @@ describe("GetCurrentWalletInstanceStatusHandler", () => {
       {
         batchPatch: () => TE.left(new Error("not implemented")),
         get: () => TE.left(new Error("not implemented")),
-        getAllByUserId: () => TE.left(new Error("not implemented")),
         getLastByUserId: () => TE.left(new ServiceUnavailableError("foo")),
+        getValidByUserIdExcludingOne: () =>
+          TE.left(new Error("not implemented")),
         insert: () => TE.left(new Error("not implemented")),
       };
     const handler = GetCurrentWalletInstanceStatusHandler({