+++ infra/resources/prod/locals.tf
@@ -2,6 +2,7 @@ locals {
   prefix         = "io"
   env_short      = "p"
   location_short = "itn"
+  domain         = "wallet"
   # the project on which the resources will be created
   # it's the prefix of any resource name
   # it includes the choosen location
@@ -53,6 +54,10 @@ locals {
         key_vault_secret_name = "PidIssuerApiClientPrivateKey"
       },
       {
+        name                  = "SlackStatusChannelWebhook"
+        key_vault_secret_name = "SlackStatusChannelWebhook"
+      },
+      {
         name                  = "AllowedDeveloperUsers"
         key_vault_secret_name = "AllowedDeveloperUsers"
       },
@@ -72,6 +77,10 @@ locals {
       {
         name                  = "AppInsightsConnectionString"
         key_vault_secret_name = "AppInsightsConnectionString"
+      },
+      {
+        name                  = "StorageConnectionString"
+        key_vault_secret_name = module.storage_accounts.wallet.connection_string_secret_name
       }
     ]
   }