+++ apps/io-wallet-support-func/CHANGELOG.md
@@ -1,5 +1,12 @@
 # io-wallet-support-func
 
+## 1.3.2
+
+### Patch Changes
+
+- Updated dependencies
+  - io-wallet-common@1.1.2
+
 ## 1.3.1
 
 ### Patch Changes